requirements.txt
aiodns==3.1.1
aiosmtpd==1.4.4.post2
altgraph==0.17.2
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
atpublic==4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.14
djangorestframework==3.15.2
dnspython==2.5.0
exceptiongroup==1.2.0
filelock==3.13.1
future==0.18.2
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
googlesearch-python==1.2.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
lxml==5.1.0
macholib==1.15.2
numpy==1.26.4
openai==1.37.1
outcome==1.3.0.post0
pandas==2.2.0
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
rsa==4.9
selenium==4.17.2
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
tldextract==5.1.1
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
verify-email==2.4.3
wsproto==1.2.0
